refact: rename wrapper
