+++ packages/designmanual/stories/LanguageWrapper/index.js
@@ -9,5 +9,4 @@
 import LanguageWrapper, { LanguageContext } from './LanguageWrapper';
 
 export { default as LanguageSelector } from './LanguageSelector';
-export { default as LanguageSummary } from './LanguageSummary';
 export { LanguageWrapper, LanguageContext };
